+++ kusto/doc.go
@@ -66,6 +66,8 @@ In this case, the StatementBuilder and QueryParameters structs should be used.
 	statement := kql.NewStatementBuilder("systemNodes | where CollectionTime == time and NodeId == id")
 	params :=  kql.NewStatementQueryParameters().AddDateTime("time", dt).AddInt("id", 1)
 	iter, err := client.Query(ctx, "database", statement, params)
+	fmt.Println(fmt.Sprintf("%s\n%s", params.ToDeclarationString(), statement.String()))
+	// Will print declare query_parameters(time:datetime, id:int)\nsystemNodes | where CollectionTime == time and NodeId == id
 	if err != nil {
 		panic("add error handling")
 	}