+++ kusto/kql/identifier.go
@@ -2,19 +2,19 @@ package kql
 
 import "fmt"
 
-func (b *statementBuilder) AddDatabase(database string) Builder {
+func (b *Builder) AddDatabase(database string) *Builder {
 	return b.addBase(stringConstant(fmt.Sprintf("%s(%s)", "database", QuoteString(database, false))))
 }
 
-func (b *statementBuilder) AddTable(table string) Builder {
+func (b *Builder) AddTable(table string) *Builder {
 	return b.addBase(stringConstant(NormalizeName(table)))
 }
 
-func (b *statementBuilder) AddColumn(column string) Builder {
+func (b *Builder) AddColumn(column string) *Builder {
 	return b.addBase(stringConstant(NormalizeName(column)))
 }
 
-func (b *statementBuilder) AddFunction(function string) Builder {
+func (b *Builder) AddFunction(function string) *Builder {
 	return b.addBase(stringConstant(NormalizeName(function)))
 }