+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 ### Added
-* Added statementBuilder for safe building of KQL statements from variables without use of 'Unsafe' mode.
-* Simpler handling of query parameters using StatementQueryParameters. 
+* Added new builder struct for safe building of KQL statements from variables without use of 'Unsafe' mode.
+* Simpler handling of query parameters using new parameters struct. 
 
 ## [0.11.1] - 2023-03-01
 ### Changed